reorganize project directories to track remaining work efficiently
