+++ src/components/LocaleSelector.svelte
