+++ src/old_components/games/setup/LogViewer.svelte
