+++ src/old_components/games/setup/Progress.svelte
