+++ src/old_components/games/setup/ProgressIcon.svelte
