+++ src/old_components/games/setup/Requirements.svelte
