--- src/splash/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en" class="dark">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/src/css/fonts.css" />
-    <script type="module" src="./splash.ts"></script>
-    <title>Splash</title>
-  </head>
-
-  <body class="m-0 overflow-hidden">
-    <div id="app" class="h-screen bg-gray-900"></div>
-  </body>
-</html>
